backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

# RAG Components
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# LLM Gateway
portkey-ai>=1.0.0

# Document Processing
pymupdf>=1.23.0
python-docx>=1.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0

